Automatic turret targeting
This works by encoding a small lookup table which maps "Target Y"
values from the vision system to "Shooter Speed" and "Cowl Angle"
values, and linearly interpolates between the two closest. The
data for this was figured out empirically by shooting a bunch of
balls at the target from various positions.

Also in this change: some tweaks to improve the reliability of
the ballevator and add a 'Back Button' (mapped to the back button)
to drive it in reverse in case a ball gets stuck.